Convert functions decorated with Provides to singletons
